+++ packages/react-querybuilder/src/components/RuleGroup.test.tsx
@@ -90,7 +90,7 @@ describe('onCombinatorChange', () => {
       container.querySelector(`.${sc.combinators}`)!,
       'any_combinator_value'
     );
-    expect(onPropChange).toHaveBeenCalledWith('combinator', 'any_combinator_value', [0]);
+    expect(onPropChange).toHaveBeenCalledWith('combinator', 'any_combinator_value', [0], undefined);
   });
 });
 
@@ -100,7 +100,7 @@ describe('onNotToggleChange', () => {
     render(<RuleGroup {...getRuleGroupProps({ showNotToggle: true }, { onPropChange })} />);
 
     await user.click(screen.getByLabelText('Not'));
-    expect(onPropChange).toHaveBeenCalledWith('not', true, [0]);
+    expect(onPropChange).toHaveBeenCalledWith('not', true, [0], undefined);
   });
 });
 
@@ -145,7 +145,46 @@ describe('cloneGroup', () => {
     render(<RuleGroup {...getRuleGroupProps({ showCloneButtons: true }, { moveRule })} />);
 
     await user.click(screen.getByText(t.cloneRuleGroup.label));
-    expect(moveRule).toHaveBeenCalledWith([0], [1], true);
+    expect(moveRule).toHaveBeenCalledWith([0], [1], true, undefined);
+  });
+});
+
+describe('ungroup', () => {
+  it('calls ungroupRuleGroup from the schema with expected values', async () => {
+    const ungroupRuleGroup = vi.fn();
+    render(
+      <RuleGroup {...getRuleGroupProps({ showUngroupButtons: true }, { ungroupRuleGroup })} />
+    );
+
+    await user.click(screen.getByText(t.ungroupRuleGroup.label));
+    expect(ungroupRuleGroup).toHaveBeenCalledWith([0], undefined);
+  });
+
+  it('does not call ungroupRuleGroup when disabled', async () => {
+    const ungroupRuleGroup = vi.fn();
+    render(
+      <RuleGroup
+        {...getRuleGroupProps({ showUngroupButtons: true }, { ungroupRuleGroup })}
+        disabled
+      />
+    );
+
+    await user.click(screen.getByText(t.ungroupRuleGroup.label));
+    expect(ungroupRuleGroup).not.toHaveBeenCalled();
+  });
+});
+
+describe('showUngroupButtons', () => {
+  it('does not display the ungroup button by default', () => {
+    const { container } = render(
+      <RuleGroup {...getRuleGroupProps({ showUngroupButtons: false })} />
+    );
+    expect(container.querySelectorAll(`.${sc.ungroup}`)).toHaveLength(0);
+  });
+
+  it('has the correct classNames', () => {
+    render(<RuleGroup {...getRuleGroupProps({ showUngroupButtons: true })} />);
+    expect(screen.getByTestId(TestID.ungroup)).toHaveClass(sc.ungroup, 'custom-ungroup-class');
   });
 });
 
@@ -177,20 +216,20 @@ describe('shiftRuleUp/Down', () => {
     rerender(<RuleGroup {...getRuleGroupProps({ showShiftActions: true }, { moveRule })} />);
 
     await user.click(screen.getByText(t.shiftActionUp.label));
-    expect(moveRule).toHaveBeenLastCalledWith([0], 'up', false);
+    expect(moveRule).toHaveBeenLastCalledWith([0], 'up', false, undefined);
 
     await user.click(screen.getByText(t.shiftActionDown.label));
-    expect(moveRule).toHaveBeenLastCalledWith([0], 'down', false);
+    expect(moveRule).toHaveBeenLastCalledWith([0], 'down', false, undefined);
 
     await user.keyboard('{Alt>}');
     await user.click(screen.getByText(t.shiftActionUp.label));
     await user.keyboard('{/Alt}');
-    expect(moveRule).toHaveBeenLastCalledWith([0], 'up', true);
+    expect(moveRule).toHaveBeenLastCalledWith([0], 'up', true, undefined);
 
     await user.keyboard('{Alt>}');
     await user.click(screen.getByText(t.shiftActionDown.label));
     await user.keyboard('{/Alt}');
-    expect(moveRule).toHaveBeenLastCalledWith([0], 'down', true);
+    expect(moveRule).toHaveBeenLastCalledWith([0], 'down', true, undefined);
   });
 });
 
@@ -311,7 +350,7 @@ describe('independent combinators', () => {
     );
 
     await user.selectOptions(screen.getByTitle(t.combinators.title), [screen.getByText('OR')]);
-    expect(onPropChange).toHaveBeenCalledWith('combinator', 'or', [0, 1]);
+    expect(onPropChange).toHaveBeenCalledWith('combinator', 'or', [0, 1], undefined);
   });
 
   it('clones independent combinator groups', async () => {
@@ -326,7 +365,7 @@ describe('independent combinators', () => {
     );
 
     await user.click(screen.getByText(t.cloneRuleGroup.label));
-    expect(moveRule).toHaveBeenCalledWith([0], [1], true);
+    expect(moveRule).toHaveBeenCalledWith([0], [1], true, undefined);
   });
 });
 
@@ -439,7 +478,7 @@ describe('lock buttons', () => {
     render(<RuleGroup {...getRuleGroupProps({ showLockButtons: true }, { onPropChange })} />);
 
     await user.click(screen.getByTestId(TestID.lockGroup));
-    expect(onPropChange).toHaveBeenCalledWith('disabled', true, [0]);
+    expect(onPropChange).toHaveBeenCalledWith('disabled', true, [0], undefined);
   });
 
   it('unsets the disabled property', async () => {
@@ -449,7 +488,7 @@ describe('lock buttons', () => {
     );
 
     await user.click(screen.getByTestId(TestID.lockGroup));
-    expect(onPropChange).toHaveBeenCalledWith('disabled', false, [0]);
+    expect(onPropChange).toHaveBeenCalledWith('disabled', false, [0], undefined);
   });
 });