+++ website/src/pages/demo/_constants/types.ts
@@ -13,6 +13,7 @@ export type DemoOption =
   | 'showShiftActions'
   | 'showUndoRedo'
   | 'showCloneButtons'
+  | 'showUngroupButtons'
   | 'showLockButtons'
   | 'showMuteButtons'
   | 'resetOnFieldChange'